html/semantics/embedded-content-0/media-elements/readyState_property_exists.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.readyState - existence test</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-networkstate">spec reference</a></p>
  <audio id="a">
  </audio>
  <video id="v">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 assert_true(
  "readyState" in document.getElementById("a"),
  "audioElement.readyState property exists");
}, "audio.readyState - existence test");

test(function() {
 assert_true(
  "readyState" in document.getElementById("v"),
  "videoElement.readyState property exists");
}, "video.readyState - existence test");
  </script>
 </body>
</html>
